Finished sort feature.

+++ app/src/main/java/employee/domain/Country.java
@@ -16,24 +16,29 @@ public class Country implements Serializable{
 	@Id
 	@GeneratedValue
 	private int id;
+	
 	@Column(nullable = false)
 	private String name;
 	
 	public Country(){}
-	
-	public long getId() {
+
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+
+	public void setId(int id) {
 		this.id = id;
 	}
+
 	public String getName() {
 		return name;
 	}
+
 	public void setName(String name) {
 		this.name = name;
 	}
 	
+
